Extracted basic arithmetic to FieldValue
