+++ src/ferris/sql/execution/engine.rs
@@ -128,24 +128,18 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 */
 
 use super::aggregation::AggregateFunctions;
-use super::expression::{ArithmeticOperations, BuiltinFunctions, ExpressionEvaluator};
+use super::expression::ExpressionEvaluator;
 use super::internal::{
-    ExecutionMessage, ExecutionState, GroupByState, HeaderMutation, HeaderOperation,
-    QueryExecution, WindowState,
+    ExecutionMessage, ExecutionState, GroupByState, HeaderMutation, QueryExecution, WindowState,
 };
 use super::types::{FieldValue, StreamRecord};
-use super::utils::{FieldValueConverter, TimeExtractor};
+use super::utils::FieldValueConverter;
 use crate::ferris::serialization::{InternalValue, SerializationFormat};
 use crate::ferris::sql::ast::{
-    Expr, JoinClause, JoinType, LiteralValue, OverClause, SelectField, StreamSource, StreamingQuery,
-    WindowSpec,
+    Expr, JoinClause, JoinType, LiteralValue, SelectField, StreamSource, StreamingQuery,
 };
 use crate::ferris::sql::error::SqlError;
-use chrono::{DateTime, NaiveDate, NaiveDateTime};
-use log::warn;
-use rust_decimal::Decimal;
 use std::collections::HashMap;
-use std::str::FromStr;
 use std::sync::Arc;
 use tokio::sync::mpsc;
 // Processor imports for Phase 5B integration
@@ -792,921 +786,9 @@ impl StreamExecutionEngine {
         }
     }
 
-    // EXPRESSION EVALUATION FUNCTIONS
-    // =============================================================================
-
-    #[doc(hidden)]
-    pub fn add_values(
-        &self,
-        left: &FieldValue,
-        right: &FieldValue,
-    ) -> Result<FieldValue, SqlError> {
-        ArithmeticOperations::add_values(left, right)
-    }
-
-    #[doc(hidden)]
-    pub fn subtract_values(
-        &self,
-        left: &FieldValue,
-        right: &FieldValue,
-    ) -> Result<FieldValue, SqlError> {
-        ArithmeticOperations::subtract_values(left, right)
-    }
-
-    #[doc(hidden)]
-    pub fn multiply_values(
-        &self,
-        left: &FieldValue,
-        right: &FieldValue,
-    ) -> Result<FieldValue, SqlError> {
-        ArithmeticOperations::multiply_values(left, right)
-    }
-
-    #[doc(hidden)]
-    pub fn divide_values(
-        &self,
-        left: &FieldValue,
-        right: &FieldValue,
-    ) -> Result<FieldValue, SqlError> {
-        ArithmeticOperations::divide_values(left, right)
-    }
-
     // =============================================================================
     // WINDOW FUNCTIONS
     // =============================================================================
-
-    /// Evaluate window functions like LAG, LEAD, ROW_NUMBER with OVER clause
-    fn evaluate_window_function(
-        &mut self,
-        function_name: &str,
-        args: &[Expr],
-        _over_clause: &OverClause,
-        record: &StreamRecord,
-        window_buffer: &mut Vec<StreamRecord>,
-    ) -> Result<FieldValue, SqlError> {
-        // Validate function name is not empty
-        if function_name.trim().is_empty() {
-            return Err(SqlError::ExecutionError {
-                message: "Window function name cannot be empty".to_string(),
-                query: Some(format!("Window function: {}", function_name)),
-            });
-        }
-
-        match function_name.to_uppercase().as_str() {
-            "LAG" => {
-                // Validate argument count
-                if args.is_empty() {
-                    return Err(SqlError::ExecutionError {
-                        message: "LAG function requires at least 1 argument (expression)"
-                            .to_string(),
-                        query: Some(format!("LAG({})", if args.is_empty() { "" } else { "..." })),
-                    });
-                }
-                if args.len() > 3 {
-                    return Err(SqlError::ExecutionError {
-                        message: format!(
-                            "LAG function accepts at most 3 arguments (expression, offset, default_value), but {} were provided",
-                            args.len()
-                        ),
-                        query: Some("LAG(expression, [offset], [default_value])".to_string()),
-                    });
-                }
-
-                // Parse offset (default is 1)
-                let offset = if args.len() >= 2 {
-                    match ExpressionEvaluator::evaluate_expression_value(&args[1], record)? {
-                        FieldValue::Integer(n) => {
-                            if n < 0 {
-                                return Err(SqlError::ExecutionError {
-                                    message: format!("LAG offset must be non-negative, got {}", n),
-                                    query: Some(format!("LAG(expression, {})", n)),
-                                });
-                            }
-                            if n > i32::MAX as i64 {
-                                return Err(SqlError::ExecutionError {
-                                    message: format!(
-                                        "LAG offset {} exceeds maximum allowed value {}",
-                                        n,
-                                        i32::MAX
-                                    ),
-                                    query: Some("LAG offset too large".to_string()),
-                                });
-                            }
-                            n as usize
-                        }
-                        FieldValue::Null => {
-                            return Err(SqlError::ExecutionError {
-                                message: "LAG offset cannot be NULL".to_string(),
-                                query: Some("LAG(expression, NULL)".to_string()),
-                            });
-                        }
-                        other => {
-                            return Err(SqlError::ExecutionError {
-                                message: format!(
-                                    "LAG offset must be an integer, got {}",
-                                    other.type_name()
-                                ),
-                                query: Some(format!(
-                                    "LAG(expression, {})",
-                                    other.type_name().to_lowercase()
-                                )),
-                            });
-                        }
-                    }
-                } else {
-                    1
-                };
-
-                // Parse default value (if provided)
-                let default_value = if args.len() == 3 {
-                    Some(ExpressionEvaluator::evaluate_expression_value(
-                        &args[2], record,
-                    )?)
-                } else {
-                    None
-                };
-
-                // Look back in the window buffer
-                if offset == 0 {
-                    // Offset 0 means current record
-                    ExpressionEvaluator::evaluate_expression_value(&args[0], record)
-                } else if window_buffer.len() >= offset {
-                    let lag_record = &window_buffer[window_buffer.len() - offset];
-                    ExpressionEvaluator::evaluate_expression_value(&args[0], lag_record)
-                } else {
-                    // Not enough records in buffer, return default or NULL
-                    Ok(default_value.unwrap_or(FieldValue::Null))
-                }
-            }
-            "LEAD" => {
-                // Validate argument count
-                if args.is_empty() {
-                    return Err(SqlError::ExecutionError {
-                        message: "LEAD function requires at least 1 argument (expression)"
-                            .to_string(),
-                        query: Some("LEAD(expression, [offset], [default_value])".to_string()),
-                    });
-                }
-                if args.len() > 3 {
-                    return Err(SqlError::ExecutionError {
-                        message: format!(
-                            "LEAD function accepts at most 3 arguments (expression, offset, default_value), but {} were provided",
-                            args.len()
-                        ),
-                        query: Some("LEAD(expression, [offset], [default_value])".to_string()),
-                    });
-                }
-
-                // Parse offset (default is 1)
-                let offset = if args.len() >= 2 {
-                    match ExpressionEvaluator::evaluate_expression_value(&args[1], record)? {
-                        FieldValue::Integer(n) => {
-                            if n < 0 {
-                                return Err(SqlError::ExecutionError {
-                                    message: format!("LEAD offset must be non-negative, got {}", n),
-                                    query: Some(format!("LEAD(expression, {})", n)),
-                                });
-                            }
-                            if n > i32::MAX as i64 {
-                                return Err(SqlError::ExecutionError {
-                                    message: format!(
-                                        "LEAD offset {} exceeds maximum allowed value {}",
-                                        n,
-                                        i32::MAX
-                                    ),
-                                    query: Some("LEAD offset too large".to_string()),
-                                });
-                            }
-                            n as usize
-                        }
-                        FieldValue::Null => {
-                            return Err(SqlError::ExecutionError {
-                                message: "LEAD offset cannot be NULL".to_string(),
-                                query: Some("LEAD(expression, NULL)".to_string()),
-                            });
-                        }
-                        other => {
-                            return Err(SqlError::ExecutionError {
-                                message: format!(
-                                    "LEAD offset must be an integer, got {}",
-                                    other.type_name()
-                                ),
-                                query: Some(format!(
-                                    "LEAD(expression, {})",
-                                    other.type_name().to_lowercase()
-                                )),
-                            });
-                        }
-                    }
-                } else {
-                    1
-                };
-
-                // Parse default value (if provided)
-                let default_value = if args.len() == 3 {
-                    Some(ExpressionEvaluator::evaluate_expression_value(
-                        &args[2], record,
-                    )?)
-                } else {
-                    None
-                };
-
-                // For streaming LEAD, we cannot look forward in most cases
-                if offset > 0 {
-                    warn!(
-                        "LEAD window function: cannot look forward {} records in streaming data - returning default value or NULL",
-                        offset
-                    );
-                    Ok(default_value.unwrap_or(FieldValue::Null))
-                } else {
-                    // Offset 0 means current record
-                    ExpressionEvaluator::evaluate_expression_value(&args[0], record)
-                }
-            }
-            "ROW_NUMBER" => {
-                if !args.is_empty() {
-                    return Err(SqlError::ExecutionError {
-                        message: format!(
-                            "ROW_NUMBER function takes no arguments, but {} were provided",
-                            args.len()
-                        ),
-                        query: Some(format!("ROW_NUMBER({} arguments)", args.len())),
-                    });
-                }
-
-                // ROW_NUMBER() OVER (...) - returns current position in partition
-                // For streaming, this is the position in the current window buffer + 1
-                let row_number = window_buffer.len() + 1;
-                if row_number > i64::MAX as usize {
-                    return Err(SqlError::ExecutionError {
-                        message: format!(
-                            "ROW_NUMBER overflow: row number {} exceeds maximum value",
-                            row_number
-                        ),
-                        query: Some("ROW_NUMBER() OVER (...)".to_string()),
-                    });
-                }
-                Ok(FieldValue::Integer(row_number as i64))
-            }
-            "RANK" | "DENSE_RANK" => {
-                if !args.is_empty() {
-                    return Err(SqlError::ExecutionError {
-                        message: format!(
-                            "{} function takes no arguments, but {} were provided",
-                            function_name.to_uppercase(),
-                            args.len()
-                        ),
-                        query: Some(format!(
-                            "{}({} arguments)",
-                            function_name.to_uppercase(),
-                            args.len()
-                        )),
-                    });
-                }
-
-                // For streaming without proper partitioning, RANK and DENSE_RANK behave like ROW_NUMBER
-                warn!(
-                    "{} window function: returning ROW_NUMBER behavior for streaming context",
-                    function_name.to_uppercase()
-                );
-                let rank = window_buffer.len() + 1;
-                if rank > i64::MAX as usize {
-                    return Err(SqlError::ExecutionError {
-                        message: format!(
-                            "{} overflow: rank {} exceeds maximum value",
-                            function_name.to_uppercase(),
-                            rank
-                        ),
-                        query: Some(format!("{}() OVER (...)", function_name.to_uppercase())),
-                    });
-                }
-                Ok(FieldValue::Integer(rank as i64))
-            }
-            "FIRST_VALUE" => {
-                if args.len() != 1 {
-                    return Err(SqlError::ExecutionError {
-                        message: format!(
-                            "FIRST_VALUE function requires exactly 1 argument (expression), but {} were provided",
-                            args.len()
-                        ),
-                        query: Some("FIRST_VALUE(expression)".to_string()),
-                    });
-                }
-
-                // Return the value from the first record in the window buffer
-                if !window_buffer.is_empty() {
-                    ExpressionEvaluator::evaluate_expression_value(&args[0], &window_buffer[0])
-                } else {
-                    // If buffer is empty, evaluate against current record
-                    ExpressionEvaluator::evaluate_expression_value(&args[0], record)
-                }
-            }
-            "LAST_VALUE" => {
-                if args.len() != 1 {
-                    return Err(SqlError::ExecutionError {
-                        message: format!(
-                            "LAST_VALUE function requires exactly 1 argument (expression), but {} were provided",
-                            args.len()
-                        ),
-                        query: Some("LAST_VALUE(expression)".to_string()),
-                    });
-                }
-
-                // Return the value from the last record in the window buffer (current record)
-                ExpressionEvaluator::evaluate_expression_value(&args[0], record)
-            }
-            "NTH_VALUE" => {
-                if args.len() != 2 {
-                    return Err(SqlError::ExecutionError {
-                        message: format!(
-                            "NTH_VALUE function requires exactly 2 arguments (expression, n), but {} were provided",
-                            args.len()
-                        ),
-                        query: Some("NTH_VALUE(expression, n)".to_string()),
-                    });
-                }
-
-                // Parse the nth position
-                let nth = match ExpressionEvaluator::evaluate_expression_value(&args[1], record)? {
-                    FieldValue::Integer(n) => {
-                        if n <= 0 {
-                            return Err(SqlError::ExecutionError {
-                                message: format!("NTH_VALUE position must be positive, got {}", n),
-                                query: Some(format!("NTH_VALUE(expression, {})", n)),
-                            });
-                        }
-                        n as usize
-                    }
-                    FieldValue::Null => {
-                        return Err(SqlError::ExecutionError {
-                            message: "NTH_VALUE position cannot be NULL".to_string(),
-                            query: Some("NTH_VALUE(expression, NULL)".to_string()),
-                        });
-                    }
-                    other => {
-                        return Err(SqlError::ExecutionError {
-                            message: format!(
-                                "NTH_VALUE position must be an integer, got {}",
-                                other.type_name()
-                            ),
-                            query: Some(format!(
-                                "NTH_VALUE(expression, {})",
-                                other.type_name().to_lowercase()
-                            )),
-                        });
-                    }
-                };
-
-                // Get the nth record from the window buffer (1-indexed)
-                let total_records = window_buffer.len() + 1; // +1 for current record
-                if nth <= total_records {
-                    if nth <= window_buffer.len() {
-                        // nth record is in the buffer
-                        ExpressionEvaluator::evaluate_expression_value(
-                            &args[0],
-                            &window_buffer[nth - 1],
-                        )
-                    } else {
-                        // nth record is the current record
-                        ExpressionEvaluator::evaluate_expression_value(&args[0], record)
-                    }
-                } else {
-                    // nth record doesn't exist
-                    Ok(FieldValue::Null)
-                }
-            }
-            "PERCENT_RANK" => {
-                if !args.is_empty() {
-                    return Err(SqlError::ExecutionError {
-                        message: format!(
-                            "PERCENT_RANK function takes no arguments, but {} were provided",
-                            args.len()
-                        ),
-                        query: Some(format!("PERCENT_RANK({} arguments)", args.len())),
-                    });
-                }
-
-                // PERCENT_RANK() = (rank - 1) / (total_rows - 1)
-                // For streaming, we use current position in buffer
-                let current_rank = window_buffer.len() + 1;
-                let total_rows = current_rank; // In streaming, we only know current position
-
-                if total_rows <= 1 {
-                    Ok(FieldValue::Float(0.0))
-                } else {
-                    let percent_rank = (current_rank - 1) as f64 / (total_rows - 1) as f64;
-                    Ok(FieldValue::Float(percent_rank))
-                }
-            }
-            "CUME_DIST" => {
-                if !args.is_empty() {
-                    return Err(SqlError::ExecutionError {
-                        message: format!(
-                            "CUME_DIST function takes no arguments, but {} were provided",
-                            args.len()
-                        ),
-                        query: Some(format!("CUME_DIST({} arguments)", args.len())),
-                    });
-                }
-
-                // CUME_DIST() = number_of_rows_with_values_<=_current_row / total_rows
-                // For streaming, we approximate this as current_position / total_known_rows
-                let current_position = window_buffer.len() + 1;
-                let total_known_rows = current_position;
-
-                let cume_dist = current_position as f64 / total_known_rows as f64;
-                Ok(FieldValue::Float(cume_dist))
-            }
-            "NTILE" => {
-                if args.len() != 1 {
-                    return Err(SqlError::ExecutionError {
-                        message: format!(
-                            "NTILE function requires exactly 1 argument (n), but {} were provided",
-                            args.len()
-                        ),
-                        query: Some("NTILE(n)".to_string()),
-                    });
-                }
-
-                // Parse the number of tiles
-                let tiles = match ExpressionEvaluator::evaluate_expression_value(&args[0], record)?
-                {
-                    FieldValue::Integer(n) => {
-                        if n <= 0 {
-                            return Err(SqlError::ExecutionError {
-                                message: format!("NTILE tiles count must be positive, got {}", n),
-                                query: Some(format!("NTILE({})", n)),
-                            });
-                        }
-                        n
-                    }
-                    FieldValue::Null => {
-                        return Err(SqlError::ExecutionError {
-                            message: "NTILE tiles count cannot be NULL".to_string(),
-                            query: Some("NTILE(NULL)".to_string()),
-                        });
-                    }
-                    other => {
-                        return Err(SqlError::ExecutionError {
-                            message: format!(
-                                "NTILE tiles count must be an integer, got {}",
-                                other.type_name()
-                            ),
-                            query: Some(format!("NTILE({})", other.type_name().to_lowercase())),
-                        });
-                    }
-                };
-
-                // Calculate which tile the current row belongs to
-                let current_row = window_buffer.len() + 1;
-                let total_rows = current_row; // In streaming, we only know current position
-
-                // Calculate tile number (1-indexed)
-                let rows_per_tile = (total_rows as f64 / tiles as f64).ceil() as i64;
-                let tile_number = ((current_row - 1) as i64 / rows_per_tile) + 1;
-                let tile_number = tile_number.min(tiles); // Ensure we don't exceed max tiles
-
-                Ok(FieldValue::Integer(tile_number))
-            }
-            other => Err(SqlError::ExecutionError {
-                message: format!(
-                    "Unsupported window function: '{}'. Supported window functions are: LAG, LEAD, ROW_NUMBER, RANK, DENSE_RANK, FIRST_VALUE, LAST_VALUE, NTH_VALUE, PERCENT_RANK, CUME_DIST, NTILE",
-                    other
-                ),
-                query: Some(format!("{}(...) OVER (...)", other)),
-            }),
-        }
-    }
-
-    fn evaluate_function_with_mutations(
-        &self,
-        name: &str,
-        args: &[Expr],
-        record: &StreamRecord,
-        header_mutations: &mut Vec<HeaderMutation>,
-    ) -> Result<FieldValue, SqlError> {
-        match name.to_uppercase().as_str() {
-            // Handle header mutation functions locally
-            "SET_HEADER" => {
-                if args.len() != 2 {
-                    return Err(SqlError::ExecutionError {
-                        message: "SET_HEADER requires exactly two arguments (key, value)"
-                            .to_string(),
-                        query: None,
-                    });
-                }
-                let key_value = ExpressionEvaluator::evaluate_expression_value(&args[0], record)?;
-                let value_field = ExpressionEvaluator::evaluate_expression_value(&args[1], record)?;
-
-                // Convert key to string representation
-                let key = match key_value {
-                    FieldValue::String(s) => s,
-                    FieldValue::Integer(i) => i.to_string(),
-                    FieldValue::Float(f) => f.to_string(),
-                    FieldValue::Boolean(b) => b.to_string(),
-                    FieldValue::Null => "null".to_string(),
-                    _ => format!("{:?}", key_value),
-                };
-
-                // Convert value to string representation
-                let value = match value_field {
-                    FieldValue::String(s) => s,
-                    FieldValue::Integer(i) => i.to_string(),
-                    FieldValue::Float(f) => f.to_string(),
-                    FieldValue::Boolean(b) => b.to_string(),
-                    FieldValue::Null => "null".to_string(),
-                    _ => format!("{:?}", value_field),
-                };
-
-                // Record the header mutation
-                header_mutations.push(HeaderMutation {
-                    operation: HeaderOperation::Set,
-                    key: key.clone(),
-                    value: Some(value.clone()),
-                });
-
-                // Return the value that was set
-                Ok(FieldValue::String(value))
-            }
-            "REMOVE_HEADER" => {
-                if args.len() != 1 {
-                    return Err(SqlError::ExecutionError {
-                        message: "REMOVE_HEADER requires exactly one argument (key)".to_string(),
-                        query: None,
-                    });
-                }
-                let key_value = ExpressionEvaluator::evaluate_expression_value(&args[0], record)?;
-
-                // Convert key to string representation
-                let key = match key_value {
-                    FieldValue::String(s) => s,
-                    FieldValue::Integer(i) => i.to_string(),
-                    FieldValue::Float(f) => f.to_string(),
-                    FieldValue::Boolean(b) => b.to_string(),
-                    FieldValue::Null => "null".to_string(),
-                    _ => format!("{:?}", key_value),
-                };
-
-                // Get the existing header value before removing it
-                let existing_value = record.headers.get(&key).cloned();
-
-                // Record the header mutation
-                header_mutations.push(HeaderMutation {
-                    operation: HeaderOperation::Remove,
-                    key: key.clone(),
-                    value: None,
-                });
-
-                // Return the value that was removed (or NULL if it didn't exist)
-                match existing_value {
-                    Some(value) => Ok(FieldValue::String(value)),
-                    None => Ok(FieldValue::Null),
-                }
-            }
-            // Delegate all other functions to the extracted module
-            _ => {
-                let func_expr = Expr::Function {
-                    name: name.to_string(),
-                    args: args.to_vec(),
-                };
-                BuiltinFunctions::evaluate_function(&func_expr, record)
-            }
-        }
-    }
-
-    #[doc(hidden)]
-    pub fn cast_value(&self, value: FieldValue, target_type: &str) -> Result<FieldValue, SqlError> {
-        match target_type {
-            "INTEGER" | "INT" => match value {
-                FieldValue::Integer(i) => Ok(FieldValue::Integer(i)),
-                FieldValue::Float(f) => Ok(FieldValue::Integer(f as i64)),
-                FieldValue::String(s) => s.parse::<i64>().map(FieldValue::Integer).map_err(|_| {
-                    SqlError::ExecutionError {
-                        message: format!("Cannot cast '{}' to INTEGER", s),
-                        query: None,
-                    }
-                }),
-                FieldValue::Boolean(b) => Ok(FieldValue::Integer(if b { 1 } else { 0 })),
-                FieldValue::Decimal(d) => {
-                    // Convert decimal to integer, truncating fractional part
-                    let int_part = d.trunc();
-                    match int_part.to_string().parse::<i64>() {
-                        Ok(i) => Ok(FieldValue::Integer(i)),
-                        Err(_) => Err(SqlError::ExecutionError {
-                            message: format!("Cannot cast DECIMAL {} to INTEGER", d),
-                            query: None,
-                        }),
-                    }
-                }
-                FieldValue::Null => Ok(FieldValue::Null),
-                FieldValue::Date(_)
-                | FieldValue::Timestamp(_)
-                | FieldValue::Array(_)
-                | FieldValue::Map(_)
-                | FieldValue::Struct(_)
-                | FieldValue::Interval { .. } => Err(SqlError::ExecutionError {
-                    message: format!("Cannot cast {} to INTEGER", value.type_name()),
-                    query: None,
-                }),
-            },
-            "FLOAT" | "DOUBLE" => match value {
-                FieldValue::Integer(i) => Ok(FieldValue::Float(i as f64)),
-                FieldValue::Float(f) => Ok(FieldValue::Float(f)),
-                FieldValue::String(s) => {
-                    s.parse::<f64>()
-                        .map(FieldValue::Float)
-                        .map_err(|_| SqlError::ExecutionError {
-                            message: format!("Cannot cast '{}' to FLOAT", s),
-                            query: None,
-                        })
-                }
-                FieldValue::Boolean(b) => Ok(FieldValue::Float(if b { 1.0 } else { 0.0 })),
-                FieldValue::Decimal(d) => {
-                    // Convert decimal to float
-                    match d.to_string().parse::<f64>() {
-                        Ok(f) => Ok(FieldValue::Float(f)),
-                        Err(_) => Err(SqlError::ExecutionError {
-                            message: format!("Cannot cast DECIMAL {} to FLOAT", d),
-                            query: None,
-                        }),
-                    }
-                }
-                FieldValue::Null => Ok(FieldValue::Null),
-                FieldValue::Date(_)
-                | FieldValue::Timestamp(_)
-                | FieldValue::Array(_)
-                | FieldValue::Map(_)
-                | FieldValue::Struct(_)
-                | FieldValue::Interval { .. } => Err(SqlError::ExecutionError {
-                    message: format!("Cannot cast {} to FLOAT", value.type_name()),
-                    query: None,
-                }),
-            },
-            "STRING" | "VARCHAR" | "TEXT" => match value {
-                FieldValue::Integer(i) => Ok(FieldValue::String(i.to_string())),
-                FieldValue::Float(f) => Ok(FieldValue::String(f.to_string())),
-                FieldValue::String(s) => Ok(FieldValue::String(s)),
-                FieldValue::Boolean(b) => Ok(FieldValue::String(b.to_string())),
-                FieldValue::Null => Ok(FieldValue::String("NULL".to_string())),
-                FieldValue::Date(_)
-                | FieldValue::Timestamp(_)
-                | FieldValue::Decimal(_)
-                | FieldValue::Array(_)
-                | FieldValue::Map(_)
-                | FieldValue::Struct(_)
-                | FieldValue::Interval { .. } => Ok(FieldValue::String(value.to_display_string())),
-            },
-            "BOOLEAN" | "BOOL" => match value {
-                FieldValue::Integer(i) => Ok(FieldValue::Boolean(i != 0)),
-                FieldValue::Float(f) => Ok(FieldValue::Boolean(f != 0.0)),
-                FieldValue::String(s) => match s.to_uppercase().as_str() {
-                    "TRUE" | "T" | "1" => Ok(FieldValue::Boolean(true)),
-                    "FALSE" | "F" | "0" => Ok(FieldValue::Boolean(false)),
-                    _ => Err(SqlError::ExecutionError {
-                        message: format!("Cannot cast '{}' to BOOLEAN", s),
-                        query: None,
-                    }),
-                },
-                FieldValue::Boolean(b) => Ok(FieldValue::Boolean(b)),
-                FieldValue::Null => Ok(FieldValue::Null),
-                FieldValue::Date(_)
-                | FieldValue::Timestamp(_)
-                | FieldValue::Decimal(_)
-                | FieldValue::Array(_)
-                | FieldValue::Map(_)
-                | FieldValue::Struct(_)
-                | FieldValue::Interval { .. } => Err(SqlError::ExecutionError {
-                    message: format!("Cannot cast {} to BOOLEAN", value.type_name()),
-                    query: None,
-                }),
-            },
-            "DATE" => match value {
-                FieldValue::Date(d) => Ok(FieldValue::Date(d)),
-                FieldValue::String(s) => NaiveDate::parse_from_str(&s, "%Y-%m-%d")
-                    .or_else(|_| NaiveDate::parse_from_str(&s, "%Y/%m/%d"))
-                    .or_else(|_| NaiveDate::parse_from_str(&s, "%m/%d/%Y"))
-                    .or_else(|_| NaiveDate::parse_from_str(&s, "%d-%m-%Y"))
-                    .map(FieldValue::Date)
-                    .map_err(|_| SqlError::ExecutionError {
-                        message: format!(
-                            "Cannot cast '{}' to DATE. Expected format: YYYY-MM-DD",
-                            s
-                        ),
-                        query: None,
-                    }),
-                FieldValue::Timestamp(ts) => Ok(FieldValue::Date(ts.date())),
-                FieldValue::Null => Ok(FieldValue::Null),
-                _ => Err(SqlError::ExecutionError {
-                    message: format!("Cannot cast {} to DATE", value.type_name()),
-                    query: None,
-                }),
-            },
-            "TIMESTAMP" | "DATETIME" => match value {
-                FieldValue::Timestamp(ts) => Ok(FieldValue::Timestamp(ts)),
-                FieldValue::Date(d) => Ok(FieldValue::Timestamp(d.and_hms_opt(0, 0, 0).unwrap())),
-                FieldValue::String(s) => {
-                    // Try various timestamp formats
-                    NaiveDateTime::parse_from_str(&s, "%Y-%m-%d %H:%M:%S")
-                        .or_else(|_| NaiveDateTime::parse_from_str(&s, "%Y-%m-%d %H:%M:%S%.3f"))
-                        .or_else(|_| NaiveDateTime::parse_from_str(&s, "%Y-%m-%dT%H:%M:%S"))
-                        .or_else(|_| NaiveDateTime::parse_from_str(&s, "%Y-%m-%dT%H:%M:%S%.3f"))
-                        .or_else(|_| NaiveDateTime::parse_from_str(&s, "%Y/%m/%d %H:%M:%S"))
-                        .or_else(|_| {
-                            // Try parsing as date only and add time
-                            NaiveDate::parse_from_str(&s, "%Y-%m-%d")
-                                .map(|d| d.and_hms_opt(0, 0, 0).unwrap())
-                        })
-                        .map(FieldValue::Timestamp)
-                        .map_err(|_| SqlError::ExecutionError {
-                            message: format!("Cannot cast '{}' to TIMESTAMP. Expected format: YYYY-MM-DD HH:MM:SS", s),
-                            query: None,
-                        })
-                }
-                FieldValue::Integer(i) => {
-                    // Treat as Unix timestamp (seconds)
-                    let dt =
-                        DateTime::from_timestamp(i, 0).ok_or_else(|| SqlError::ExecutionError {
-                            message: format!("Invalid Unix timestamp: {}", i),
-                            query: None,
-                        })?;
-                    Ok(FieldValue::Timestamp(dt.naive_utc()))
-                }
-                FieldValue::Null => Ok(FieldValue::Null),
-                _ => Err(SqlError::ExecutionError {
-                    message: format!("Cannot cast {} to TIMESTAMP", value.type_name()),
-                    query: None,
-                }),
-            },
-            "DECIMAL" | "NUMERIC" => {
-                match value {
-                    FieldValue::Decimal(d) => Ok(FieldValue::Decimal(d)),
-                    FieldValue::Integer(i) => Ok(FieldValue::Decimal(Decimal::from(i))),
-                    FieldValue::Float(f) => Decimal::from_str(&f.to_string())
-                        .map(FieldValue::Decimal)
-                        .map_err(|_| SqlError::ExecutionError {
-                            message: format!("Cannot cast float {} to DECIMAL", f),
-                            query: None,
-                        }),
-                    FieldValue::String(s) => Decimal::from_str(&s)
-                        .map(FieldValue::Decimal)
-                        .map_err(|_| SqlError::ExecutionError {
-                            message: format!("Cannot cast '{}' to DECIMAL", s),
-                            query: None,
-                        }),
-                    FieldValue::Boolean(b) => Ok(FieldValue::Decimal(if b {
-                        Decimal::ONE
-                    } else {
-                        Decimal::ZERO
-                    })),
-                    FieldValue::Null => Ok(FieldValue::Null),
-                    _ => Err(SqlError::ExecutionError {
-                        message: format!("Cannot cast {} to DECIMAL", value.type_name()),
-                        query: None,
-                    }),
-                }
-            }
-            _ => Err(SqlError::ExecutionError {
-                message: format!("Unsupported cast target type: {}", target_type),
-                query: None,
-            }),
-        }
-    }
-
-
-    /// Process JOIN clauses and combine records
-    fn process_joins(
-        &mut self,
-        left_record: &StreamRecord,
-        join_clauses: &[JoinClause],
-    ) -> Result<StreamRecord, SqlError> {
-        let mut result_record = left_record.clone();
-
-        for join_clause in join_clauses {
-            // Get right record using JoinContext implementation
-            let join_context = JoinContext;
-            let right_record_opt =
-                join_context.get_right_record(&join_clause.right_source, &join_clause.window)?;
-
-            match join_clause.join_type {
-                JoinType::Inner => {
-                    if let Some(right_record) = right_record_opt {
-                        let combined_record = self.combine_records(
-                            &result_record,
-                            &right_record,
-                            &join_clause.right_alias,
-                        )?;
-                        if ExpressionEvaluator::evaluate_expression(
-                            &join_clause.condition,
-                            &combined_record,
-                        )? {
-                            result_record = combined_record;
-                        } else {
-                            // INNER JOIN: if condition fails, no result
-                            return Err(SqlError::ExecutionError {
-                                message: "INNER JOIN condition not met".to_string(),
-                                query: None,
-                            });
-                        }
-                    } else {
-                        // INNER JOIN: if no right record, no result
-                        return Err(SqlError::ExecutionError {
-                            message: "No matching record for INNER JOIN".to_string(),
-                            query: None,
-                        });
-                    }
-                }
-                JoinType::Left => {
-                    if let Some(right_record) = right_record_opt {
-                        let combined_record = self.combine_records(
-                            &result_record,
-                            &right_record,
-                            &join_clause.right_alias,
-                        )?;
-                        if ExpressionEvaluator::evaluate_expression(
-                            &join_clause.condition,
-                            &combined_record,
-                        )? {
-                            result_record = combined_record;
-                        } else {
-                            // LEFT JOIN: if condition fails, use left record with NULL right fields
-                            result_record = self.combine_records_with_nulls(
-                                &result_record,
-                                &join_clause.right_alias,
-                                true,
-                            )?;
-                        }
-                    } else {
-                        // LEFT JOIN: if no right record, use left record with NULL right fields
-                        result_record = self.combine_records_with_nulls(
-                            &result_record,
-                            &join_clause.right_alias,
-                            true,
-                        )?;
-                    }
-                }
-                JoinType::Right => {
-                    if let Some(right_record) = right_record_opt {
-                        let combined_record = self.combine_records(
-                            &result_record,
-                            &right_record,
-                            &join_clause.right_alias,
-                        )?;
-                        if ExpressionEvaluator::evaluate_expression(
-                            &join_clause.condition,
-                            &combined_record,
-                        )? {
-                            result_record = combined_record;
-                        } else {
-                            // RIGHT JOIN: if condition fails, use right record with NULL left fields
-                            result_record = self.combine_records_with_nulls(
-                                &right_record,
-                                &join_clause.right_alias,
-                                false,
-                            )?;
-                        }
-                    } else {
-                        // RIGHT JOIN: if no right record, no result (this scenario is rare in stream processing)
-                        return Err(SqlError::ExecutionError {
-                            message: "No right record for RIGHT JOIN".to_string(),
-                            query: None,
-                        });
-                    }
-                }
-                JoinType::FullOuter => {
-                    if let Some(right_record) = right_record_opt {
-                        let combined_record = self.combine_records(
-                            &result_record,
-                            &right_record,
-                            &join_clause.right_alias,
-                        )?;
-                        if ExpressionEvaluator::evaluate_expression(
-                            &join_clause.condition,
-                            &combined_record,
-                        )? {
-                            result_record = combined_record;
-                        } else {
-                            // FULL OUTER JOIN: if condition fails, could return both records with NULLs
-                            // For simplicity, we'll return left record with NULL right fields
-                            result_record = self.combine_records_with_nulls(
-                                &result_record,
-                                &join_clause.right_alias,
-                                true,
-                            )?;
-                        }
-                    } else {
-                        // FULL OUTER JOIN: if no right record, use left record with NULL right fields
-                        result_record = self.combine_records_with_nulls(
-                            &result_record,
-                            &join_clause.right_alias,
-                            true,
-                        )?;
-                    }
-                }
-            }
-        }
-
-        Ok(result_record)
-    }
 
     /// Optimized stream-table JOIN processing
     /// Tables are materialized views with key-based lookups, streams are continuous data
@@ -1720,207 +802,6 @@ impl StreamExecutionEngine {
     ) -> Result<Option<StreamRecord>, SqlError> {
         Err(SqlError::ExecutionError {
             message: "Stream-table JOIN operations are not yet implemented. This requires materialized table state, key-based lookups, and proper streaming join semantics.".to_string(),
-            query: None,
-        })
-    }
-
-    /// Combine left and right records with NULL values for missing side
-    fn combine_records_with_nulls(
-        &self,
-        base_record: &StreamRecord,
-        right_alias: &Option<String>,
-        is_left_join: bool, // true for LEFT JOIN, false for RIGHT JOIN
-    ) -> Result<StreamRecord, SqlError> {
-        let mut combined_fields = base_record.fields.clone();
-
-        // Add NULL fields for the missing side
-        if is_left_join {
-            // LEFT JOIN: add NULL right fields
-            let right_prefix = if let Some(alias) = right_alias {
-                format!("{}.", alias)
-            } else {
-                "right_".to_string()
-            };
-
-            // Add common right-side fields as NULL
-            combined_fields.insert(format!("{}id", right_prefix), FieldValue::Null);
-            combined_fields.insert(format!("{}name", right_prefix), FieldValue::Null);
-            combined_fields.insert(format!("{}value", right_prefix), FieldValue::Null);
-        } else {
-            // RIGHT JOIN: add NULL left fields (preserve original left field names as NULL)
-            // This is more complex as we'd need to know the left schema
-            // For now, we'll just keep the right record as-is
-        }
-
-        Ok(StreamRecord {
-            fields: combined_fields,
-            timestamp: base_record.timestamp,
-            offset: base_record.offset,
-            partition: base_record.partition,
-            headers: base_record.headers.clone(),
-        })
-    }
-
-    /// Combine left and right records for JOIN processing
-    fn combine_records(
-        &self,
-        left: &StreamRecord,
-        right: &StreamRecord,
-        right_alias: &Option<String>,
-    ) -> Result<StreamRecord, SqlError> {
-        let mut combined_fields = left.fields.clone();
-
-        // Add right side fields, with optional alias prefix
-        for (key, value) in &right.fields {
-            let field_name = if let Some(alias) = right_alias {
-                format!("{}.{}", alias, key)
-            } else {
-                // If no alias, prefix with "right_" to avoid conflicts
-                format!("right_{}", key)
-            };
-            combined_fields.insert(field_name, value.clone());
-        }
-
-        Ok(StreamRecord {
-            fields: combined_fields,
-            timestamp: left.timestamp, // Use left side timestamp
-            offset: left.offset,
-            partition: left.partition,
-            headers: left.headers.clone(),
-        })
-    }
-
-    // =============================================================================
-    // HEADER MUTATION FUNCTIONS
-    // =============================================================================
-
-    fn collect_header_mutations_from_fields(
-        &self,
-        fields: &[SelectField],
-        record: &StreamRecord,
-        header_mutations: &mut Vec<HeaderMutation>,
-    ) -> Result<(), SqlError> {
-        for field in fields {
-            if let SelectField::Expression { expr, .. } = field {
-                self.collect_header_mutations_from_expr(expr, record, header_mutations)?;
-            }
-        }
-        Ok(())
-    }
-
-    fn collect_header_mutations_from_expr(
-        &self,
-        expr: &Expr,
-        record: &StreamRecord,
-        header_mutations: &mut Vec<HeaderMutation>,
-    ) -> Result<(), SqlError> {
-        match expr {
-            Expr::Function { name, args } => {
-                let function_name = name.to_uppercase();
-                if function_name == "SET_HEADER" || function_name == "REMOVE_HEADER" {
-                    // Evaluate the function with header mutations
-                    self.evaluate_function_with_mutations(
-                        &function_name,
-                        args,
-                        record,
-                        header_mutations,
-                    )?;
-                }
-
-                // Also check arguments for nested header function calls
-                for arg in args {
-                    self.collect_header_mutations_from_expr(arg, record, header_mutations)?;
-                }
-            }
-            Expr::BinaryOp { left, right, .. } => {
-                self.collect_header_mutations_from_expr(left, record, header_mutations)?;
-                self.collect_header_mutations_from_expr(right, record, header_mutations)?;
-            }
-            Expr::UnaryOp { expr, .. } => {
-                self.collect_header_mutations_from_expr(expr, record, header_mutations)?;
-            }
-            Expr::Case {
-                when_clauses,
-                else_clause,
-            } => {
-                for (condition, result) in when_clauses {
-                    self.collect_header_mutations_from_expr(condition, record, header_mutations)?;
-                    self.collect_header_mutations_from_expr(result, record, header_mutations)?;
-                }
-                if let Some(else_expr) = else_clause {
-                    self.collect_header_mutations_from_expr(else_expr, record, header_mutations)?;
-                }
-            }
-            Expr::WindowFunction { args, .. } => {
-                for arg in args {
-                    self.collect_header_mutations_from_expr(arg, record, header_mutations)?;
-                }
-            }
-            _ => {} // Other expression types don't contain function calls
-        }
-        Ok(())
-    }
-
-    /// Evaluate subquery expressions (IN, EXISTS, scalar subqueries, etc.)
-
-    // =============================================================================
-    // WINDOW PROCESSING IMPLEMENTATION
-    // =============================================================================
-
-    /// Check if window should emit based on window specification and current time
-    fn should_emit_window(
-        &self,
-        window_state: &WindowState,
-        current_time: i64,
-        _time_column: Option<&str>,
-    ) -> bool {
-        match &window_state.window_spec {
-            WindowSpec::Tumbling { size, .. } => {
-                let window_size_ms = size.as_millis() as i64;
-                let current_window_start =
-                    TimeExtractor::align_to_window_boundary(current_time, window_size_ms);
-
-                // For tumbling windows, only emit when:
-                // 1. This is the first window (last_emit == 0) AND we have the first record past window size
-                // 2. We've crossed into a new window boundary
-
-                if window_state.last_emit == 0 {
-                    // First window - only emit if we have records past the first window boundary
-                    current_time >= window_size_ms
-                } else {
-                    // Subsequent windows - emit when we cross window boundaries
-                    current_window_start > window_state.last_emit
-                }
-            }
-            WindowSpec::Sliding { advance, .. } => {
-                let advance_ms = advance.as_millis() as i64;
-                current_time >= window_state.last_emit + advance_ms
-            }
-            WindowSpec::Session { gap, .. } => {
-                let gap_ms = gap.as_millis() as i64;
-                // Session window emits when gap timeout is exceeded
-                current_time >= window_state.last_emit + gap_ms
-            }
-        }
-    }
-
-    /// Find field by suffix patterns
-    fn find_field_by_suffix(
-        &self,
-        result_record: &StreamRecord,
-        suffixes: &[&str],
-    ) -> Result<String, SqlError> {
-        for field_name in result_record.fields.keys() {
-            for suffix in suffixes {
-                // Check both exact match and suffix match
-                if field_name == suffix || field_name.ends_with(suffix) {
-                    return Ok(field_name.clone());
-                }
-            }
-        }
-
-        Err(SqlError::ExecutionError {
-            message: format!("No field found with suffixes: {:?}", suffixes),
             query: None,
         })
     }
