+++ src/ferris/sql/execution/expression/mod.rs
@@ -3,7 +3,6 @@
 //! This module contains all the logic for evaluating SQL expressions including:
 //! - Expression evaluation (boolean and value expressions)
 //! - Built-in function implementations (math, string, date functions)
-//! - Arithmetic operations (add, subtract, multiply, divide)
 //! - Window functions (LAG, LEAD, ROW_NUMBER, RANK, etc.)
 //!
 //! The expression evaluation system supports:
@@ -14,14 +13,12 @@
 //! - Type coercion and casting
 //! - NULL handling according to SQL semantics
 
-pub mod arithmetic;
 pub mod evaluator;
 pub mod functions;
 pub mod subquery_executor;
 pub mod window_functions;
 
 // Re-export the main API
-pub use arithmetic::ArithmeticOperations;
 pub use evaluator::ExpressionEvaluator;
 pub use functions::BuiltinFunctions;
 pub use subquery_executor::{SubqueryExecutor, evaluate_subquery_with_executor};
